01-intro-llms/examples/requirements.txt
openai>=2.15.0
python-dotenv>=1.2.1
requests>=2.32.5
pydantic>=2.12.5
anthropic>=0.77.0
google-genai>=1.61.0